requirements.txt
groq
python-dotenv
ipython
jupyter
requests
agentic_patterns
